+++ pyobjc-core/Modules/objc/test/testbndl.m
@@ -518,8 +518,8 @@ static char* g_charps[] = {
 - (char*)charpArg:(char*)arg
 {
 static 	char buf[1024];
-	int len = strlen(arg);
-	int i;
+	size_t len = strlen(arg);
+	size_t i;
 
 	for (i = 0; i < len; i++) {
 		buf[len - i - 1] = arg[i];
@@ -758,7 +758,7 @@ static 	char buf[1024];
 - (char*)passInCharp:(char**)arg
 {
 	/* Yes this is leaking, but we're only testing method calling */
-	int len = strlen(*arg);
+	size_t len = strlen(*arg);
 	char* res = malloc(len * 2 + 1);
 	char* p;
 	char* q;
@@ -780,7 +780,7 @@ static 	char buf[1024];
 - (void)passInOutCharp:(char**)arg
 {
 	/* Yes this is leaking, but we're only testing method calling */
-	int len = strlen(*arg);
+	size_t len = strlen(*arg);
 	char* res = malloc(len * 2 + 1);
 	char* p;
 	char* q;
